improve error/warn handling, rearrange duty filterers
