+++ analyzer-v2/duties/proposer.go
@@ -9,6 +9,7 @@ import (
 	"github.com/attestantio/go-eth2-client/spec/phase0"
 
 	"github.com/ssvlabs/ssv-pulse/analyzer-v2/internal/environment"
+	"github.com/ssvlabs/ssv-pulse/analyzer-v2/internal/helper"
 	"github.com/ssvlabs/ssv-pulse/internal/analyzer/parser"
 )
 
@@ -22,6 +23,7 @@ var dutyStepsProposer = []string{
 	"got partial RANDAO signatures",
 	"got pre consensus quorum",
 	"reconstructed partial RANDAO signatures",
+	"waited out proposer delay",
 	"got beacon block proposal",
 	// TODO - this should eventually be replaced by the next step ("fetched attestation data from CL")
 	"starting QBFT instance",
@@ -38,7 +40,6 @@ var dutyStepsProposer = []string{
 	"QBFT instance is decided",
 	"broadcasted post consensus partial signature message",
 	"got post consensus quorum",
-	"waited out proposer delay",
 	"submitting block proposal",
 	"successfully submitted block proposal",
 	"successfully finished duty processing",
@@ -73,24 +74,24 @@ func (s *Proposer) Analyze(logFilePath string, targetSlot phase0.Slot) error {
 		line := scanner.Text()
 		lineNumber++
 
-		if !relevantForSlot(line, targetSlot) {
-			continue
-		}
-		if !relevantForProposerDuty(line) {
+		lineIsRelevant := func() bool {
+			if !relevantForSlot(line, targetSlot) {
+				return false
+			}
+
+			if containsUnexpectedError(line) || containsUnexpectedWarn(line) {
+				return true
+			}
+
+			return relevantForProposerDuty(line)
+		}()
+
+		if !lineIsRelevant {
 			continue
 		}
 
-		lineIsRelevant := false
-		for _, dutyStep := range dutyStepsProposer {
-			if strings.Contains(line, dutyStep) {
-				lineIsRelevant = true
-				break
-			}
-		}
-		if lineIsRelevant {
-			if err := s.logWithTimeIntoSlot(logger, line, lineNumber, targetSlot); err != nil {
-				return err
-			}
+		if err := s.logWithTimeIntoSlot(logger, line, lineNumber, targetSlot); err != nil {
+			return err
 		}
 	}
 	err = scanner.Err()
@@ -117,4 +118,24 @@ func (s *Proposer) logWithTimeIntoSlot(logger *slog.Logger, line string, lineNum
 	logger.Info(timeIntoSlotStr + " " + trimmedLine)
 
 	return nil
+}
+
+func relevantForProposerDuty(line string) bool {
+	// TODO - gotta filter by validator-pubkey sometimes as well ?
+	//const vPubkey = "903dff3e6a2615754803e58e320d206056535c354c1b650793b0c14c00017de4fc341b25869928a83a3bcaa45f943379"
+	//if !strings.Contains(line, vPubkey) {
+	//	return false
+	//}
+
+	if !helper.ContainsCaseInsensitive(line, "proposer") {
+		return false
+	}
+
+	for _, dutyStep := range dutyStepsProposer {
+		if strings.Contains(line, dutyStep) {
+			return true
+		}
+	}
+
+	return false
 }
